requirements.txt
alembic==1.14.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
hubspot-api-client==8.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.7
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.21.0
python-dotenv
gunicorn==20.1.0
uvicorn==0.22.0
Werkzeug==2.2.2